LayoutTests/fast/forms/color/display-none-input-color-chooser-shown-expected.txt

Pick a color
PASS: Color picker is shown after clicking label for display:none color input
